Break up party resource file to resolve merge conflict
